backend/requirements.txt
# FastAPI Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
asyncpg>=0.29.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2

# WebSocket
websockets>=12.0

# Pydantic
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Logging
loguru>=0.7.2

# CORS
fastapi-cors>=0.0.6

# Utils
python-dotenv>=1.0.0
aiofiles>=23.2.1
